IDatabase now requires implementing Iterable<IEntry>
Implemented Iterable<IEntry> in ListingDB

+++ Project/src/UseCase/Listing/ListingDB.java
@@ -1,14 +1,15 @@
 package UseCase.Listing;
 
+import Entities.IEntry;
 import Entities.Listing.Listing;
 import Entities.Listing.ListingType;
 import UseCase.IDatabase;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
+import java.util.function.Consumer;
 
-public class ListingDB implements IDatabase, Map<ListingType, ArrayList<Listing>> {
+public class ListingDB implements IDatabase{
 
     private final HashMap<ListingType, ArrayList<Listing>> listingDB;
 
@@ -17,7 +18,7 @@ public class ListingDB implements IDatabase, Map<ListingType, ArrayList<Listing>
 
         for (Listing listing:
              listings) {
-            addListing(listing);
+            addEntry(listing);
         }
     }
 
@@ -28,43 +29,64 @@ public class ListingDB implements IDatabase, Map<ListingType, ArrayList<Listing>
     /**
      * Adds a listing to the database if the listing does not already exist in the database
      *
-     * @param listing - The listing to be added to the database
+     * @param entry - The entry to be added to the database
      * @return true of the listing was successfully added to the database. False otherwise
      */
-    public boolean addListing(Listing listing) {
-        ListingType type = listing.getListingType();
+    public boolean addEntry(IEntry entry) {
+        if (entry instanceof Listing){
+            Listing listing = (Listing) entry;
+            ListingType type = listing.getListingType();
 
-        if (!listingDB.containsKey(type)){
-            listingDB.put(type, new ArrayList<>());
-        }
+            if (!listingDB.containsKey(type)){
+                listingDB.put(type, new ArrayList<>());
+            }
 
-        if (getListingIndex(listing) != -1) {
-            return false;
-        } else {
-            listingDB.get(type).add(listing);
-            return true;
+            if (getIndex(listing) == -1) {
+                listingDB.get(type).add(listing);
+                return true;
+            }
         }
+        return false;
     }
 
     /**
-     * Update the listing by removing the old listing and replacing it with a new one.
+     * Update the listing by removing the old listing and replacing it with a new one. If no old listing exists, it
+     * simply adds the listing to the database;
      * returns true if listing was successfully updated.
-     * returns false if no such listing had existed in the database.
+     * returns false if listing was not updated for any reason.
      *
      * // This is done this way for thread safety. If listing was being updated as, for example, a driver was attempting
      * // to display it, it would cause some concurrentAccess related exceptions. Until I figure out how that can be
      * // correctly done, the listing would have to be replaced entirely.
      *
-     * @param listing
+     * @param entry the entry to be added
      * @return
      */
-    public boolean updateListing(Listing listing){
-        int index = getListingIndex(listing);
-        if (index != -1){
+    @Override
+    public boolean updateEntry(IEntry entry){
+        if (entry instanceof Listing){
+            Listing listing = (Listing) entry;
             ArrayList<Listing> db = listingDB.get(listing.getListingType());
-            db.remove(index);
+            int index = getIndex(listing);
+            if (index != -1){
+                db.remove(index);
+            }
             db.add(listing);
             return true;
+        }
+        return false;
+    }
+
+    /**
+     * returns true if an Entry is contained in this database, otherwise returns false.
+     *
+     * @param entry
+     * @return
+     */
+    @Override
+    public boolean contains(IEntry entry) {
+        if (entry instanceof Listing){
+            return (getIndex((Listing) entry) != -1);
         }
         return false;
     }
@@ -77,7 +99,7 @@ public class ListingDB implements IDatabase, Map<ListingType, ArrayList<Listing>
      * @param listing
      * @return
      */
-    private int getListingIndex(Listing listing){
+    private int getIndex(Listing listing){
         ListingType type = listing.getListingType();
         ArrayList<Listing> db = listingDB.get(type);
         for (int i = 0; i < db.size(); i++) {
@@ -88,67 +110,66 @@ public class ListingDB implements IDatabase, Map<ListingType, ArrayList<Listing>
         return -1;
     }
 
+    /**
+     * Removes the provided entry from the database.
+     *
+     *
+     * @param entry
+     * @return
+     */
+    @Override
+    public boolean removeEntry(IEntry entry) {
+        if (entry instanceof Listing){
+            Listing listing = (Listing) entry;
+            int index = getIndex(listing);
+            if (index != -1){
+                listingDB.get(listing.getListingType()).remove(index);
+                return true;
+            }
+        }
+        return false;
+    }
+
     @Override
     public int size() {
-        return listingDB.size();
+
+        int size = 0;
+        for (ListingType key:
+                listingDB.keySet()) {
+            size += listingDB.get(key).size();
+        }
+        return size;
     }
 
-    @Override
-    public boolean isEmpty() {
-        return listingDB.isEmpty();
-    }
-
-    @Override
-    public boolean containsKey(Object key) {
-        return listingDB.containsKey(key);
-    }
-
-    @Override
-    public boolean containsValue(Object value) {
-        return listingDB.containsValue(value);
-    }
-
-    @Override
-    public ArrayList<Listing> get(Object key) {
-        return listingDB.get(key);
-    }
-
-    @Nullable
-    @Override
-    public ArrayList<Listing> put(ListingType key, ArrayList<Listing> value) {
-        return listingDB.put(key, value);
-    }
-
-    @Override
-    public ArrayList<Listing> remove(Object key) {
-        return listingDB.remove(key);
-    }
-
-    @Override
-    public void putAll(@NotNull Map<? extends ListingType, ? extends ArrayList<Listing>> m) {
-        listingDB.putAll(m);
-    }
-
-    @Override
-    public void clear() {
-        listingDB.clear();
-    }
 
     @NotNull
     @Override
-    public Set<ListingType> keySet() {
-        return listingDB.keySet();
+    public Iterator<IEntry> iterator() {
+        return new ListingDBIterator(this.listingDB);
     }
 
-    @NotNull
-    @Override
-    public Collection<ArrayList<Listing>> values() {
-        return listingDB.values();
-    }
 
-    @NotNull
-    @Override
-    public Set<Entry<ListingType, ArrayList<Listing>>> entrySet() {
-        return listingDB.entrySet();
+    static class ListingDBIterator implements Iterator<IEntry>{
+
+        private final Iterator<Listing> toIterate;
+
+        public ListingDBIterator(HashMap<ListingType, ArrayList<Listing>> listingMap){
+            ArrayList<Listing> totalList = new ArrayList<>();
+            for (ListingType key:
+                 listingMap.keySet()) {
+                totalList.addAll(listingMap.get(key));
+            }
+            toIterate = totalList.iterator();
+        }
+
+        @Override
+        public boolean hasNext() {
+            return toIterate.hasNext();
+        }
+
+        @Override
+        public IEntry next() {
+            return toIterate.next();
+        }
     }
 }
